--- a/ident.go
+++ b/ident.go
@@ -38,9 +38,9 @@ func toParts(s string) []string {
 		return []string{strings.ToUpper(s)}
 	}
 	var prev rune
-	var x string
+	var x strings.Builder
+	x.Grow(len(s))
 	for _, c := range s {
-		cs := string(c)
 		// fmt.Println("### cs ->", cs)
 		// fmt.Println("### unicode.IsControl(c) ->", unicode.IsControl(c))
 		// fmt.Println("### unicode.IsDigit(c) ->", unicode.IsDigit(c))
@@ -58,35 +58,38 @@ func toParts(s string) []string {
 		}
 
 		if isSpace(c) {
-			parts = xappend(parts, x)
-			x = cs
+			parts = xappend(parts, x.String())
+			x.Reset()
+			x.WriteRune(c)
 			prev = c
 			continue
 		}
 
 		if unicode.IsUpper(c) && !unicode.IsUpper(prev) {
-			parts = xappend(parts, x)
-			x = cs
+			parts = xappend(parts, x.String())
+			x.Reset()
+			x.WriteRune(c)
 			prev = c
 			continue
 		}
-		if unicode.IsUpper(c) && baseAcronyms[strings.ToUpper(x)] {
-			parts = xappend(parts, x)
-			x = cs
+		if unicode.IsUpper(c) && baseAcronyms[strings.ToUpper(x.String())] {
+			parts = xappend(parts, x.String())
+			x.Reset()
+			x.WriteRune(c)
 			prev = c
 			continue
 		}
 		if unicode.IsLetter(c) || unicode.IsDigit(c) || unicode.IsPunct(c) || c == '`' {
 			prev = c
-			x += cs
+			x.WriteRune(c)
 			continue
 		}
 
-		parts = xappend(parts, x)
-		x = ""
+		parts = xappend(parts, x.String())
+		x.Reset()
 		prev = c
 	}
-	parts = xappend(parts, x)
+	parts = xappend(parts, x.String())
 
 	return parts
 }