--- a/underscore.go
+++ b/underscore.go
@@ -18,16 +18,17 @@ func Underscore(s string) string {
 //	Nice to see you! = nice_to_see_you
 //	widgetID = widget_id
 func (i Ident) Underscore() Ident {
-	var out []string
+	out := make([]string, 0, len(i.Parts))
 	for _, part := range i.Parts {
-		var x string
+		var x strings.Builder
+		x.Grow(len(part))
 		for _, c := range part {
 			if unicode.IsLetter(c) || unicode.IsDigit(c) {
-				x += string(c)
+				x.WriteRune(c)
 			}
 		}
-		if x != "" {
-			out = append(out, x)
+		if x.Len() > 0 {
+			out = append(out, x.String())
 		}
 	}
 	return New(strings.ToLower(strings.Join(out, "_")))