Was browsing the code to see why `repo` becomes `repoes` and made some quick performance improvements that made a big enougth diffrence thought worth sharing.

`New()` is 18% faster

Before:
`Benchmark_New-8   	   26959	     43947 ns/op	   12736 B/op	     665 allocs/op`
After:
`Benchmark_New-8   	   33150	     36114 ns/op	   12288 B/op	     498 allocs/op`

  
`Underscore` is 5% faster (20% including changes to New())

Before:
`Benchmark_Underscore-8   	   43320	     26377 ns/op	    8576 B/op	     446 allocs/op`
After:
`Benchmark_Underscore-8   	   44116	     24971 ns/op	    8256 B/op	     411 allocs/op`
WithNew:
`Benchmark_Underscore-8   	   57582	     20848 ns/op	    8208 B/op	     328 allocs/op`